src/requirements.txt
uvicorn
fastapi
httpx
pyjwt
python-decouple
pytest
requests